requirements.txt
pandas==2.2.2
python_dateutil==2.9.0.post0
PyYAML==6.0.2
